+++ scripts/helpers.sh
@@ -56,9 +56,11 @@ capture_pane_contents_option_on() {
 	[ "$option" == "on" ]
 }
 
-save_bash_history_option_on() {
-	local option="$(get_tmux_option "$bash_history_option" "off")"
-	[ "$option" == "on" ]
+save_shell_history_option_on() {
+	local option_shell="$(get_tmux_option "$shell_history_option" "off")"
+	local option_bash="$(get_tmux_option "$bash_history_option" "off")"
+
+	[ "$option_shell" == "on" ] || [ "$option_bash" == "on" ]
 }
 
 get_grouped_sessions() {
@@ -126,5 +128,6 @@ pane_contents_archive_file() {
 
 resurrect_history_file() {
 	local pane_id="$1"
-	echo "$(resurrect_dir)/bash_history-${pane_id}"
+	local shell_name="$2"
+	echo "$(resurrect_dir)/${shell_name}_history-${pane_id}"
 }