+++ scripts/variables.sh
@@ -37,7 +37,10 @@ pane_contents_option="@resurrect-capture-pane-contents"
 pane_contents_area_option="@resurrect-pane-contents-area"
 default_pane_contents_area="full"
 
+# Keep @resurrect-save-bash-history for backward compatibility.
+# It's effectively an alias to @resurrect-save-shell-history.
 bash_history_option="@resurrect-save-bash-history"
+shell_history_option="@resurrect-save-shell-history"
 
 # set to 'on' to ensure panes are never ever overwritten
 overwrite_option="@resurrect-never-overwrite"